Simplify spotlight and global layer handling
